fix: Actually track nickname changes
